# plugins/rest-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zonvo AI REST API
  description: |
    Welcome to the Zonvo AI REST API. Everything your AI voice agents do on the phone — calling, qualifying,
    booking, promising a callback, honouring a do-not-call request — is available to your CRM, ERP or
    in-house app through `https://app.zonvo.tech/api/v1`.

    ## What you can automate

    | Goal | Endpoints |
    |------|-----------|
    | **Trigger calls** from your own app, now or at a scheduled time | `POST /calls` (with `scheduledAt`, `dynamicVariables`) |
    | **Run bulk campaigns** | `POST /campaigns`, `/campaigns/{id}/contacts`, `/start`, `/pause` |
    | **Sync leads and call outcomes to your CRM** | `GET /leads`, `PATCH /leads/{id}`, `POST /leads/{id}/notes`, `GET /calls?outcome=`, `POST /calls/{id}/outcome` |
    | **Book, confirm or cancel appointments** the agent made | `GET /appointments`, `PATCH /appointments/{id}` |
    | **Schedule callbacks** (single or in bulk, with idempotency keys) | `POST /callbacks`, `POST /callbacks/bulk`, `DELETE /callbacks?externalRef=` |
    | **Keep a do-not-call list** in sync with your compliance system | `GET /dnd`, `POST /dnd`, `POST /dnd/import`, `GET /dnd/check` |
    | **Receive real-time webhooks** for calls, appointments, callbacks and leads | `POST /webhooks`, `GET /webhooks/events` |

    ## Integrating from a CRM in 5 steps

    1. **Create an API key** — Settings → Developer → *Create API Key*. Give it only the scopes you need
...

...

...

...

...
        ]
...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /webhooks/{id}/test:
    post:
      summary: Test Webhook
      description: |
        Send a test payload to verify your webhook endpoint is configured correctly. The test payload mimics a real event.
        
        <br />
        
        **Test Process:**
        1. Zonvo AI sends a test payload to your URL
        2. Your server should return 2xx status
        3. Response includes success/failure and timing info
        
        <br />
        
        **Troubleshooting:**
        - Ensure your endpoint is publicly accessible
        - Check that you respond within 30 seconds
        - Verify SSL certificate is valid
        
        <br />
        
        **Required Scope:** `webhooks:write`
      operationId: testWebhook
      tags: [Webhooks]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Test results with timing and response info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookTestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /appointments:
    get:
      summary: List Appointments
      description: |
        Appointments booked by your agents and Flows. `from` / `to` filter on `appointmentDate` (YYYY-MM-DD or ISO date-time).

        **Required Scope:** `calls:read`
      operationId: listAppointments
      tags: [Appointments]
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
          example: "2026-09-24"
        - name: to
          in: query
          schema:
            type: string
            format: date
          example: "2026-09-30"
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, confirmed, cancelled, completed, no_show]
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Appointments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentListResponse'
              example:
                success: true
                data:
                  - id: "b7e2f9a1-6c3d-4e5f-8a9b-0c1d2e3f4a5b"
                    contactName: "Priya Sharma"
                    contactPhone: "+919876543210"
                    contactEmail: "priya@example.com"
                    appointmentDate: "2026-09-30"
                    appointmentTime: "11:00:00"
                    duration: 30
                    serviceName: "Product demo"
                    notes: "Interested in the Growth plan"
                    status: scheduled
                    callId: "c9f1d1a2-3b4c-4d5e-8f60-1a2b3c4d5e6f"
                    googleCalendarEventId: "k3j4h5g6f7d8s9a0"
                    createdAt: "2026-09-24T06:12:41.000Z"
                meta:
                  requestId: "6b1f0c1e-3d2a-4c5b-8e9f-0a1b2c3d4e5f"
                  timestamp: "2026-09-24T08:00:00.000Z"
                  pagination: { page: 1, pageSize: 20, totalItems: 1, totalPages: 1, hasNext: false, hasPrev: false }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /appointments/{id}:
    get:
      summary: Get Appointment
      description: |
        **Required Scope:** `calls:read`
      operationId: getAppointment
      tags: [Appointments]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update Appointment Status
      description: |
        Confirm, cancel, complete or mark as no-show. When the appointment has a Google Calendar event it is
        updated best-effort — cancelling deletes the event and clears `googleCalendarEventId`. The matching webhook
        (`appointment.confirmed`, `appointment.cancelled`, `appointment.completed`, `appointment.no_show`) fires only
        when the status actually changes.

        **Required Scope:** `calls:write`
      operationId: updateAppointment
      tags: [Appointments]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppointmentRequest'
            example:
              status: confirmed
              statusReason: "Customer confirmed on WhatsApp"
      responses:
        '200':
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token (API key)
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Meta:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    PaginationMeta:
      allOf:
        - $ref: '#/components/schemas/Meta'
        - type: object
          properties:
            pagination:
              type: object
              properties:
                page:
                  type: integer
                pageSize:
                  type: integer
                totalItems:
                  type: integer
                totalPages:
                  type: integer
                hasNext:
                  type: boolean
                hasPrev:
                  type: boolean
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request parameters
            details:
              type: object
        meta:
          $ref: '#/components/schemas/Meta'
...
    WebhookTestResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            delivered:
              type: boolean
            statusCode:
              type: integer
            message:
              type: string
        meta:
          $ref: '#/components/schemas/Meta'

    # ---------------------------------------------------------------- Call outcomes
...
    Appointment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contactName:
          type: string
        contactPhone:
          type: string
        contactEmail:
          type: string
          nullable: true
        appointmentDate:
          type: string
          format: date
          example: "2026-09-30"
        appointmentTime:
          type: string
          description: Local time (HH:MM:SS) in the agent's appointment time zone
          example: "11:00:00"
        duration:
          type: integer
          description: Minutes
        serviceName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [scheduled, confirmed, cancelled, completed, no_show]
        statusReason:
          type: string
          nullable: true
        callId:
          type: string
          nullable: true
        googleCalendarEventId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AppointmentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Appointment'
        meta:
          $ref: '#/components/schemas/Meta'
    AppointmentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    UpdateAppointmentRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [scheduled, confirmed, cancelled, completed, no_show]
        statusReason:
          type: string
          maxLength: 500